requirements.txt
attrs==21.2.0
beautifulsoup4==4.9.3
chardet==4.0.0
click==8.0.1
Flask==2.0.1
iniconfig==1.1.1
isort==5.9.2
itsdangerous==2.0.1
Jinja2==3.0.1
levenshtein==0.12.0
MarkupSafe==2.0.1
numpy==1.21.0
opencv-python==4.5.2.54
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pyocr==0.8
pyparsing==2.4.7
pytest==6.2.4
soupsieve==2.2.1
toml==0.10.2
Werkzeug==2.0.1
yapf==0.31.0
